+++ lib/src/core/models/check_3ds_version/check_3ds_version_response.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../constants.dart';
+import '../../../constants.dart';
 import '../../models/base/acquiring_response.dart';
 import '../../models/enums/status.dart';