+++ lib/src/core/models/remove_card/remove_card_response.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../constants.dart';
+import '../../../constants.dart';
 import '../base/acquiring_response.dart';
 import '../enums/card_status.dart';