+++ test/tinkoff_acquiring_test.dart
@@ -1,17 +1,12 @@
 import 'package:flutter_test/flutter_test.dart';
 
 import '../test_components/crypto_test.dart' as crypto_test;
-import '../test_components/hash_test.dart' as hash_test;
-import '../test_components/sdk_single_test.dart' as single_test;
 import '../test_components/sdk_complex_test.dart' as complex_test;
+import '../test_components/sdk_single_test.dart' as single_test;
 
 Future<void> main() async {
   group('Crypto Tests', () {
     crypto_test.start();
-  });
-
-  group('Hash Tests', () {
-    hash_test.start();
   });
 
   group('SDK Single Tests', () {